Tidy tests/plugins/core/
