+++ tests/plugins/core/test_importing.py
@@ -1,10 +1,7 @@
-# pylint: disable=redefined-outer-name, unused-argument
+# pylint: disable=unused-argument
 
 import os
 import pathlib
-from typing import Dict, Type
-
-import pytest
 
 import tempor
 import tempor.plugins.core._plugin as plugin_core
@@ -43,31 +40,6 @@ class DummyPluginTempor(plugin_core.Plugin):
     pass
 
 """
-
-DUMMY_PLUGIN_CATEGORY_REGISTRY: Dict[str, Type[plugin_core.Plugin]] = dict()
-DUMMY_PLUGIN_REGISTRY: Dict[str, Type[plugin_core.Plugin]] = dict()
-
-
-@pytest.fixture
-def patch_plugins_core_module(patch_module):
-    patch_module(
-        main_module=plugin_core,
-        module_vars=[
-            (
-                plugin_core,
-                plugin_core.PLUGIN_CATEGORY_REGISTRY,
-                "PLUGIN_CATEGORY_REGISTRY",
-                DUMMY_PLUGIN_CATEGORY_REGISTRY,
-            ),
-            (
-                plugin_core,
-                plugin_core.PLUGIN_REGISTRY,
-                "PLUGIN_REGISTRY",
-                DUMMY_PLUGIN_REGISTRY,
-            ),
-        ],
-        refresh_pydantic=False,
-    )
 
 
 def test_import_plugins(tmp_path: pathlib.Path, patch_plugins_core_module):
